// apps/intelliclone/lib/database.types.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      conversations: {
        Row: {
          chatbot_id: string
          created_at: string
          id: string
          reference_id: string
          user_email: string | null
        }
        Insert: {
          chatbot_id: string
          created_at?: string
          id?: string
          reference_id: string
          user_email?: string | null
        }
        Update: {
          chatbot_id?: string
          created_at?: string
          id?: string
          reference_id?: string
          user_email?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "conversations_chatbot_id_fkey"
            columns: ["chatbot_id"]
            isOneToOne: false
            referencedRelation: "chatbots"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      messages: {
        Row: {
          chatbot_id: string
          conversation_id: string
          created_at: string
          id: number
          sender: Database["public"]["Enums"]["sender"]
          text: string
          type: Database["public"]["Enums"]["message_type"]
        }
        Insert: {
          chatbot_id: string
          conversation_id: string
          created_at?: string
          id?: number
          sender: Database["public"]["Enums"]["sender"]
          text: string
          type: Database["public"]["Enums"]["message_type"]
        }
        Update: {
          chatbot_id?: string
          conversation_id?: string
          created_at?: string
          id?: number
          sender?: Database["public"]["Enums"]["sender"]
          text?: string
          type?: Database["public"]["Enums"]["message_type"]
        }
        Relationships: [
          {
            foreignKeyName: "messages_chatbot_id_fkey"
            columns: ["chatbot_id"]
            isOneToOne: false
            referencedRelation: "chatbots"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "messages_conversation_id_fkey"
            columns: ["conversation_id"]
            isOneToOne: false
            referencedRelation: "conversations"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
// ...
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never